requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.4.1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.5.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==2.0.0
colorama==0.4.6
cryptography==41.0.4
filelock==3.12.4
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.4
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
load-dotenv==0.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
narwhals==1.8.1
numpy==2.1.1
oscrypto==1.3.0
packaging==23.2
pandas==2.2.2
pillow==10.4.0
platformdirs==3.11.0
protobuf==5.28.1
psycopg2==2.9.9
pyarrow==17.0.0
pycparser==2.21
pycryptodomex==3.19.0
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rich==13.8.1
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
smmap==5.0.1
snowflake-connector-python==3.2.1
snowflake-snowpark-python==1.8.0
sortedcontainers==2.4.0
soupsieve==2.5
streamlit==1.38.0
tenacity==8.5.0
threadpoolctl==3.5.0
toml==0.10.2
tomlkit==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2023.3
urllib3==1.26.17
watchdog==4.0.2
yarl==1.9.4
